Close #285 - Redesign loggerf.core.syntax and loggerf.cats.syntax
