+++ modules/logger-f-cats/src/main/scala-2/loggerf/cats/syntax/all.scala
@@ -1,0 +1,6 @@
+package loggerf.cats.syntax
+
+/** @author Kevin Lee
+  * @since 2022-02-20
+  */
+object all extends LogMessageSyntax with LogSyntax
