+++ modules/logger-f-cats/src/main/scala-3/loggerf/cats/syntax/LogMessageSyntax.scala
@@ -1,0 +1,9 @@
+package loggerf.cats.syntax
+
+/**
+ * @author Kevin Lee
+ * @since 2022-02-20
+ */
+trait LogMessageSyntax extends loggerf.core.syntax.LogMessageSyntax
+
+object LogMessageSyntax extends LogMessageSyntax
