+++ modules/logger-f-cats/src/main/scala-3/loggerf/cats/syntax/all.scala
@@ -1,0 +1,3 @@
+package loggerf.cats.syntax
+
+object all extends LogMessageSyntax with LogSyntax
