+++ modules/logger-f-core/src/main/scala-2/loggerf/core/syntax/all.scala
@@ -1,0 +1,6 @@
+package loggerf.core.syntax
+
+/** @author Kevin Lee
+  * @since 2022-02-20
+  */
+object all extends LogMessageSyntax with LogSyntax
