+++ modules/logger-f-core/src/main/scala-3/loggerf/core/syntax/all.scala
@@ -1,0 +1,3 @@
+package loggerf.core.syntax
+
+object all extends LogMessageSyntax with LogSyntax
