+++ modules/logger-f-core/src/main/scala/loggerf/logger/CanLog.scala
@@ -1,5 +1,7 @@
 package loggerf.logger
 
+import loggerf.Level
+
 /** @author Kevin Lee
   * @since 2020-03-28
   */
@@ -8,4 +10,15 @@ trait CanLog {
   def info(message: => String): Unit
   def warn(message: => String): Unit
   def error(message: => String): Unit
+}
+
+object CanLog {
+  implicit class GetLogger(val canLog: CanLog) extends AnyVal {
+    @inline def getLogger(level: Level): (=> String) => Unit = level match {
+      case Level.Debug => canLog.debug
+      case Level.Info => canLog.info
+      case Level.Warn => canLog.warn
+      case Level.Error => canLog.error
+    }
+  }
 }
