--- modules/logger-f-core/src/main/scala/loggerf/syntax/package.scala
@@ -1,6 +1,0 @@
-package loggerf
-
-/** @author Kevin Lee
-  * @since 2020-08-01
-  */
-package object syntax extends LogSyntax with LeveledMessageSyntax
