+++ modules/logger-f-core/src/test/scala/loggerf/future/instancesSpec.scala
@@ -14,7 +14,7 @@ import hedgehog.runner._
 import loggerf.core._
 import loggerf.future.instances.logFuture
 import loggerf.logger._
-import loggerf.syntax._
+import loggerf.core.syntax.all._
 
 import java.util.concurrent.ExecutorService
 import scala.concurrent.duration._
