+++ modules/logger-f-core/src/test/scala/loggerf/future/syntaxSpec.scala
@@ -12,9 +12,8 @@ import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
 import hedgehog.runner._
 import loggerf.core.Log
-import loggerf.future.syntax._
+import loggerf.core.syntax.all._
 import loggerf.logger._
-import loggerf.syntax._
 
 import java.util.concurrent.ExecutorService
 import scala.concurrent.duration._
