+++ modules/logger-f-monix/src/test/scala/loggerf/monix/instancesSpec.scala
@@ -9,7 +9,7 @@ import hedgehog.runner._
 import loggerf.core.Log
 import loggerf.logger.LoggerForTesting
 import loggerf.cats.instances.logF
-import loggerf.syntax._
+import loggerf.cats.syntax.all._
 import monix.eval.Task
 
 /** @author Kevin Lee
